+++ missiongen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.5.0"
+__version__ = "1.5.1"
 
 from .recipe import Recipe
 from .builder import StarterBuilder, generate